[1, 2, 3, 4 단계] Web MVC (#9)
* docs: 요구사항 정의

* refactor: 사용자 역할 추가

* feat: 토큰 발급 및 로그인 기능 구현

* docs: 요구사항 추가 정의

* feat: 로그인한 멤버가 내 정보를 조회하는 기능 추가

* refactor: 로그인한 사용자만 예약이 가능하도록 리팩터링

* feat: 자신의 예약이 아닌 경우 예약 취소 기능 추가

* refactor: 사용자 역할 분리

* refactor: 관리자 기능 보호

* feat: 예약이력 조회 기능 추가

* feat: 예약 이력, 플레이 이력 관련 기능 구현

+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -21,7 +21,7 @@ public class ReservationService {
         this.scheduleDao = scheduleDao;
     }
 
-    public Long create(ReservationRequest reservationRequest) {
+    public Long create(ReservationRequest reservationRequest, Long memberId) {
         Schedule schedule = scheduleDao.findById(reservationRequest.getScheduleId());
         if (schedule == null) {
             throw new NullPointerException();
@@ -34,6 +34,7 @@ public class ReservationService {
 
         Reservation newReservation = new Reservation(
                 schedule,
+                memberId,
                 reservationRequest.getName()
         );
 
@@ -49,12 +50,12 @@ public class ReservationService {
         return reservationDao.findAllByThemeIdAndDate(themeId, date);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id, Long memberId) {
         Reservation reservation = reservationDao.findById(id);
         if (reservation == null) {
             throw new NullPointerException();
         }
-
+        reservation.validateOwner(memberId);
         reservationDao.deleteById(id);
     }
 }
